+++ app/db/truncate.py
@@ -1,5 +1,7 @@
 from app.db.database import SessionLocal
-from app.models.models import RunDate, Task, User
+from app.models.run_date import RunDate
+from app.models.task import Task
+from app.models.user import User
 
 db = SessionLocal()
