--- app/models/models.py
@@ -1,26 +1,0 @@
-from sqlalchemy import Column, DateTime, ForeignKey, Integer, String
-
-from ..db.database import Base
-
-
-class User(Base):
-    __tablename__ = "users"
-    id = Column(Integer, primary_key=True, index=True)
-    name = Column(String, index=True)
-    email = Column(String, unique=True, index=True)
-    password = Column(String, index=True)
-
-
-class Task(Base):
-    __tablename__ = "tasks"
-    id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, ForeignKey("users.id"), index=True)
-    name = Column(String, index=True)
-    order = Column(Integer, unique=True, index=True)
-
-
-class RunDate(Base):
-    __tablename__ = "rundates"
-    id = Column(Integer, primary_key=True, index=True)
-    task_id = Column(Integer, ForeignKey("tasks.id"), index=True)
-    date = Column(DateTime, index=True)
