+++ app/models/user.py
@@ -1,0 +1,11 @@
+from sqlalchemy import Column, Integer, String
+
+from ..db.database import Base
+
+
+class User(Base):
+    __tablename__ = "users"
+    id = Column(Integer, primary_key=True, index=True)
+    name = Column(String, index=True)
+    email = Column(String, unique=True, index=True)
+    password = Column(String, index=True)
