+++ app/repositories/crud.py
@@ -2,28 +2,30 @@ import hashlib
 
 from sqlalchemy.orm import Session
 
-from app.models import models
+from app.models.run_date import RunDate
+from app.models.task import Task
+from app.models.user import User
 from app.schemas import schemas
 
 # Read
 
 
 def get_users(db: Session, skip: int = 0, limit: int = 100):
-    return db.query(models.User).offset(skip).limit(limit).all()
+    return db.query(User).offset(skip).limit(limit).all()
 
 
 def get_tasks(db: Session, skip: int = 0, limit: int = 100):
-    return db.query(models.Task).offset(skip).limit(limit).all()
+    return db.query(Task).offset(skip).limit(limit).all()
 
 
 def get_run_dates(db: Session, skip: int = 0, limit: int = 100):
-    return db.query(models.RunDate).offset(skip).limit(limit).all()
+    return db.query(RunDate).offset(skip).limit(limit).all()
 
 # Create
 
 
 def create_user(db: Session, user: schemas.UserCreate):
-    db_user = models.User(
+    db_user = User(
         name=user.name,
         email=user.email,
         password=hashlib.sha256(user.password.encode()).hexdigest()
@@ -35,7 +37,7 @@ def create_user(db: Session, user: schemas.UserCreate):
 
 
 def create_task(db: Session, task: schemas.TaskCreate):
-    db_task = models.Task(
+    db_task = Task(
         user_id=task.user_id,
         name=task.name,
         order=task.order
@@ -47,7 +49,7 @@ def create_task(db: Session, task: schemas.TaskCreate):
 
 
 def create_run_date(db: Session, run_date: schemas.RunDateCreate):
-    db_run_date = models.RunDate(
+    db_run_date = RunDate(
         task_id=run_date.task_id,
         date=run_date.date
     )
